Update to latest Tchernobog version, plus minor changes and improvements
Eaglehorn/Goreshovel typo fixed to match latest version of Tchernobog.

Added alwaysShowNormal, alwaysShowMagic, and alwaysShowRare.

Changed strictness scaling formula so that no scaling happens at strictness = 0.

For those used to the old scaling you can use oldStrictnessScaling = true. This will not be documented on the configurable variables page on the wiki as it is basically a deprecated calculation.